pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coverage==7.4.1",
    "dj-database-url>=2.1.0",
    "django==5.0.2",
    "django-cors-headers==4.3.1",
    "django-filter==23.5",
    "djangorestframework==3.14.0",
    "drf-yasg==1.21.7",
    "gunicorn==21.2.0",
    "psycopg2-binary>=2.9.9",
    "pytest==8.0.0",
    "pytest-django==4.8.0",
    "python-dotenv==1.0.1",
    "setuptools>=65.5.1",
    "whitenoise==6.6.0",
]
